+++ packages/client/src/stores/command-palette.store.ts
@@ -2,14 +2,20 @@ import { create } from "zustand";
 
 interface CommandPaletteState {
   paletteOpen: boolean;
+  shortcutsOpen: boolean;
   openPalette: () => void;
   closePalette: () => void;
   togglePalette: () => void;
+  openShortcuts: () => void;
+  closeShortcuts: () => void;
 }
 
 export const useCommandPaletteStore = create<CommandPaletteState>((set) => ({
   paletteOpen: false,
-  openPalette: () => set({ paletteOpen: true }),
+  shortcutsOpen: false,
+  openPalette: () => set({ paletteOpen: true, shortcutsOpen: false }),
   closePalette: () => set({ paletteOpen: false }),
-  togglePalette: () => set((state) => ({ paletteOpen: !state.paletteOpen })),
+  togglePalette: () => set((state) => ({ paletteOpen: !state.paletteOpen, shortcutsOpen: false })),
+  openShortcuts: () => set({ shortcutsOpen: true, paletteOpen: false }),
+  closeShortcuts: () => set({ shortcutsOpen: false }),
 }));